Rename datafu.pig.numbers to datafu.pig.random
